+++ src/mapgen_fractal.h
@@ -57,23 +57,12 @@ struct MapgenFractalParams : public MapgenSpecificParams {
 	void writeParams(Settings *settings) const;
 };
 
-class MapgenFractal : public Mapgen {
+class MapgenFractal : public MapgenBasic {
 public:
-	EmergeManager *m_emerge;
-	BiomeManager *bmgr;
-
-	int ystride;
-	int zstride_1d;
 	u16 formula;
 	bool julia;
 
-	v3s16 node_min;
-	v3s16 node_max;
-	v3s16 full_node_min;
-	v3s16 full_node_max;
-
 	u32 spflags;
-	float cave_width;
 	u16 fractal;
 	u16 iterations;
 	v3f scale;
@@ -84,16 +73,9 @@ public:
 	float julia_z;
 	float julia_w;
 	Noise *noise_seabed;
-	Noise *noise_filler_depth;
-	Noise *noise_cave1;
-	Noise *noise_cave2;
 
-	content_t c_stone;
-	content_t c_water_source;
 	content_t c_lava_source;
-	content_t c_desert_stone;
 	content_t c_ice;
-	content_t c_sandstone;
 
 	content_t c_cobble;
 	content_t c_stair_cobble;
@@ -109,9 +91,6 @@ public:
 	void calculateNoise();
 	bool getFractalAtPoint(s16 x, s16 y, s16 z);
 	s16 generateTerrain();
-	MgStoneType generateBiomes();
-	void dustTopNodes();
-	void generateCaves(s16 max_stone_y);
 };
 
 struct MapgenFactoryFractal : public MapgenFactory {
