+++ src/mapgen_v5.h
@@ -48,34 +48,15 @@ struct MapgenV5Params : public MapgenSpecificParams {
 };
 
 
-class MapgenV5 : public Mapgen {
+class MapgenV5 : public MapgenBasic {
 public:
-	EmergeManager *m_emerge;
-	BiomeManager *bmgr;
-
-	int ystride;
-	int zstride_1d;
-
-	v3s16 node_min;
-	v3s16 node_max;
-	v3s16 full_node_min;
-	v3s16 full_node_max;
-
 	u32 spflags;
-	float cave_width;
-	Noise *noise_filler_depth;
 	Noise *noise_factor;
 	Noise *noise_height;
-	Noise *noise_cave1;
-	Noise *noise_cave2;
 	Noise *noise_ground;
 
-	content_t c_stone;
-	content_t c_water_source;
 	content_t c_lava_source;
-	content_t c_desert_stone;
 	content_t c_ice;
-	content_t c_sandstone;
 
 	content_t c_cobble;
 	content_t c_stair_cobble;
@@ -90,9 +71,6 @@ public:
 	int getSpawnLevelAtPoint(v2s16 p);
 	void calculateNoise();
 	int generateBaseTerrain();
-	MgStoneType generateBiomes();
-	void generateCaves(int max_stone_y);
-	void dustTopNodes();
 };
